add hook for extension loading; default blank view background to white
